requirements.txt
boto3==1.9.119
botocore==1.12.119
dj-database-url==0.5.0
Django==2.2.2
django-autocomplete-light==3.4.1
django-ckeditor==5.7.1
django-easy-pjax==1.3.0
django-heroku==0.3.1
django-image-cropping==0.3.7
django-js-asset==1.2.2
django-reset-migrations==0.3.1
django-storages==1.7.1
django-tinymce==2.8.0
docutils==0.14
easy-thumbnails==2.6
gunicorn==19.9.0
image==1.5.27
jmespath==0.9.4
olefile==0.46
Pillow==5.4.1
psycopg2==2.7.7
python-dateutil==2.8.0
python-dotenv==0.10.1
pytz==2018.9
s3transfer==0.2.0
six==1.12.0
sqlparse==0.3.0
urllib3==1.24.1
whitenoise==4.1.2
